[FIXED JENKINS-44898] Add findResource to PluginFirstClassLoader (#2916)
* [FIXED JENKINS-44898] Add findResource to PluginFirstClassLoader

This fixes GroovyClassLoader.loadClass for a .groovy file in a plugin
with a PluginFirstClassLoader, specifically by fixing fast-loading via
the UberClassLoader.

* Move common code to AntWithFindResourceClassLoader.

* Adding testing of new PluginFirstClassLoader behavior.
